projeto_compras_estrategicas_2026/requirements.txt
pandas
matplotlib
pdfminer.six